split widgets into multiple file, rename few things, add require for some constructors
