+++ lib/entity/story.dart
@@ -3,7 +3,7 @@ class Story {
   int id;
   List<int> kids;
   int parent;
-  String text;
+  String title;
   int time;
   int score;
   String type;
@@ -13,7 +13,7 @@ class Story {
         this.id,
         this.kids,
         this.parent,
-        this.text,
+        this.title,
         this.time,
         this.score,
         this.type});
@@ -23,7 +23,7 @@ class Story {
     id = json['id'];
     kids = json['kids'] != null ? json['kids'].cast<int>() : <int>[];
     parent = json['parent'];
-    text = json['title'] != null ? json['title'] : "";
+    title = json['title'] != null ? json['title'] : "";
     time = json['time'];
     score = json['score'] != null ? json['score'] : 0;
     type = json['type'];
@@ -35,7 +35,7 @@ class Story {
     data['id'] = this.id;
     data['kids'] = this.kids;
     data['parent'] = this.parent;
-    data['text'] = this.text;
+    data['text'] = this.title;
     data['time'] = this.time;
     data['score'] = this.score;
     data['type'] = this.type;
